pyproject.toml
[project]
name = "paper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "arxiv>=2.4.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "httpx[socks]>=0.28.1",
    "ipython>=8.39.0",
    "requests>=2.32.5",
    "google-genai>=1.0.0",
]

[tool.uv]
index-url = "http://mirrors.aliyun.com/pypi/simple" 
package = true
